chore: move plh-prgroess=path component from plh_kids_kw scope to plh scope
