+++ packages/components/plh/index.ts
@@ -5,7 +5,7 @@ import { PlhActivityCheckInComponent } from "./plh-kids-kw/components/activity-c
 import { PlhCompletionModalComponent } from "./plh-kids-kw/components/completion-modal/completion-modal.component";
 import { PlhModuleDetailsHeaderComponent } from "./plh-kids-kw/components/module-details-header/module-details-header.component";
 import { PlhBottomNavigationBarComponent } from "./plh-kids-kw/components/bottom-navigation-bar/bottom-navigation-bar.component";
-import { PlhProgressPathComponent } from "./plh-kids-kw/components/plh-progress-path/plh-progress-path.component";
+import { PlhProgressPathComponent } from "./progress-path/progress-path.component";
 
 export {
   PlhParentPointCounterComponent,
