+++ packages/components/plh/progress-path/progress-path.component.scss
